apps/arena/pyproject.toml
[project]
name = "liz-arena"
version = "0.1.0"
description = "LLM Arena - Compare vision models for maintenance intake classification"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.50",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "groq>=0.11",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
